% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susologparser-package.R
\docType{package}
\name{susologparser-package}
\alias{susologparser}
\alias{susologparser-package}
\title{susologparser: Shiny Application to parse Survey Solutions Log Files}
\description{
A simple shiny application which allows to parse Survey Solutions log files, and to select specific events, for review and download. This application, may be helpful, to generate the required information for the Survey Solutions support team, when reporting issues.
}
\author{
\strong{Maintainer}: Michael Wild \email{mwild@worldbank.org}

}
\keyword{internal}
